# dokumentation/openapi.yaml
openapi: 3.1.0
info:
  title: Template Editor Service
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /workspaces/comments:
    get:
      tags:
      - Workspaces Comments
      summary: Kommentare eines Merge Requests laden
      description: Liest alle nicht-System-Kommentare aus GitLab Discussions und mappt
        sie in das interne CommentsPayload.
      operationId: getComments
      parameters:
      - name: repositoryId
        in: query
        description: GitLab Project/Repository ID
        required: true
        schema:
          type: string
          pattern: "^[0-9]+$"
        example: 12345
      - name: mrId
        in: query
        description: Merge Request IID
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
        example: 17
      - name: branch
        in: query
        description: Branch-Name
        required: true
        schema:
          type: string
          pattern: "^[A-Za-z0-9._/-]+$"
        example: feature/my-branch
      - name: version
        in: query
        description: Version/Tag des Templates (für Mapping/Resolver)
        required: true
        schema:
          type: string
          pattern: "^\\d+\\.\\d+\\.\\d+(?:[-+][A-Za-z0-9.-]+)?$"
        example: 1.2.3
      responses:
        "200":
          description: Kommentare erfolgreich geladen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
        "404":
          description: Merge Request/Repository nicht gefunden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
        "400":
          description: Bad Request – validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
        "401":
          description: Unauthorized – missing or invalid Bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentsPayload"
      security:
      - bearerAuth: []
    post:
      tags:
      - Workspaces Comments
      summary: Inline-Kommentar erstellen
      description: Erstellt eine neue GitLab Discussion an einer Dateiposition (newPath
        + newLine).
      operationId: createComment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateRequest"
        required: true
      responses:
        "201":
          description: Kommentar erstellt
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "404":
          description: Merge Request/Datei nicht gefunden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "400":
          description: Bad Request – validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "401":
          description: Unauthorized – missing or invalid Bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
      security:
      - bearerAuth: []
  /workspaces/comments/reply:
    post:
      tags:
      - Workspaces Comments
      summary: Auf Thread antworten (und optional resolven)
      description: "Erstellt eine Note in einer bestehenden Discussion. Wenn resolved=true,\
        \ wird der Thread anschließend resolved."
      operationId: reply
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyRequest"
        required: true
      responses:
        "201":
          description: Reply erstellt
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "404":
          description: Thread/MR nicht gefunden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "400":
          description: Bad Request – validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "401":
          description: Unauthorized – missing or invalid Bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
      security:
      - bearerAuth: []
...
  /workspaces/details:
    get:
      tags:
      - Workspaces
      summary: Get workspace details for a branch
      description: "Returns all workspace files (templates, metadata, changelog, notes,\
        \ …) for the given repository/branch/version."
      operationId: details
      parameters:
      - name: repositoryId
        in: query
        description: GitLab repository / project identifier
        required: true
        schema:
          type: string
          pattern: "^[0-9]+$"
        example: 1234
      - name: branch
        in: query
        description: Branch name (workspace) to open
        required: true
        schema:
          type: string
          pattern: "^[A-Za-z0-9._/-]+$"
        example: feature/new-terminology
      - name: version
        in: query
        description: Terminology package version
        required: true
        schema:
          type: string
          pattern: "^\\d+\\.\\d+\\.\\d+(?:[-+][A-Za-z0-9.-]+)?$"
        example: 1.0.3
      responses:
        "200":
          description: Workspace successfully loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetails"
        "400":
          description: Bad Request – validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceDetails"
        "401":
          description: Unauthorized – missing or invalid Bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceDetails"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceDetails"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceDetails"
      security:
      - bearerAuth: []
...
  /projects:
    get:
      tags:
      - Projects
      summary: List GitLab projects
      description: "Returns a paginated list of GitLab projects, optionally searchable\
        \ and sortable."
      operationId: listProjects
      parameters:
      - name: search
        in: query
        description: Filters projects by substring match in name
        required: false
        schema:
          type: string
          pattern: "^[\\p{L}0-9 _\\-.]*$"
        example: template
      - name: sort
        in: query
        description: Sort order of projects
        required: false
        schema:
          type: string
          enum:
          - lastModified
          - title
        example: lastModified
      - name: page
        in: query
        description: 1-based page index
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          maximum: 10000
          minimum: 1
        example: 1
      - name: pageSize
        in: query
        description: Number of items per page (1–200)
        required: false
        schema:
          type: integer
          format: int32
          default: 50
          maximum: 200
          minimum: 1
        example: 50
      responses:
        "200":
          description: Projects successfully listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectItem"
        "400":
          description: Bad Request – validation failed
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectItem"
        "401":
          description: Unauthorized – missing or invalid Bearer token
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectItem"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectItem"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectItem"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    MergeRequestRef:
      type: object
      description: Merge request that was created for the commit
      properties:
        id:
          type: integer
          format: int32
          description: Merge request IID
          example: 42
        webUrl:
          type: string
          description: GitLab web URL of the merge request
          example: https://gitlab.example.com/group/project/-/merge_requests/42
    CommentCreateRequest:
      type: object
      description: Request zum Erstellen eines Merge-Request Inline-Kommentars (GitLab
        Discussion).
      properties:
        repositoryId:
          type: string
          description: "GitLab Projekt-ID (wird als `projects/{id}` verwendet)."
          example: "12345"
          minLength: 1
          pattern: "^[0-9]+$"
        branch:
          type: string
          description: "Branch, der zur ResourceIndex-Auflösung verwendet wird."
          example: feature/my-branch
          minLength: 1
          pattern: "^[A-Za-z0-9._/-]+$"
        mrId:
          type: string
          description: Merge-Request IID als String.
          example: "17"
          minLength: 1
          pattern: "^[0-9]+$"
        version:
          type: string
          description: Version-Prefix im Repo (z.B. Ordner `<version>/...`).
          example: 1.2.3
          minLength: 1
          pattern: "^\\d+\\.\\d+\\.\\d+(?:[-+][A-Za-z0-9.-]+)?$"
        type:
          type: string
          description: "Art der kommentierten Resource. Bestimmt, wie `fileName` in\
            \ einen GitLab-Pfad gemappt wird."
          enum:
          - template
          - template_markdown
          - package_template
          - metadata_package
          - changelogs
          - package_markdown
          - update_index_url_version
          - download_conditions
          - input_file
          example: template
        fileName:
          type: string
          description: Identifier der Datei/Resource abhängig vom `type`.
          example: my.template.json
          minLength: 1
        line:
          type:
          - integer
          - "null"
          format: int32
          description: Zeilennummer in der Diff-Position. Muss gesetzt sein.
          example: 42
          minimum: 1
        body:
          type: string
          description: Kommentartext für GitLab Discussion/Note.
          example: Bitte hier die Formulierung anpassen.
          minLength: 1
          pattern: "^[^\\p{Cntrl}]+$"
      required:
      - body
      - branch
      - fileName
      - line
      - mrId
      - repositoryId
      - type
      - version
    CommentCreateResponse:
      type: object
      description: Response nach dem Erstellen eines Kommentars oder Replies. Enthält
        die ID der erzeugten GitLab Note/Discussion.
      properties:
        id:
          type: string
          description: ID der erstellten GitLab Note bzw. Discussion.
          example: "123456789"
    ReplyRequest:
      type: object
      description: Request zum Antworten auf eine bestehende GitLab Discussion (Thread).
      properties:
        repositoryId:
          type: string
          writeOnly: true
        mrId:
          type: integer
          format: int32
          description: Merge Request IID.
          example: 17
          minimum: 1
        threadId:
          type: string
          description: ID der GitLab Discussion (Thread).
          example: 61568f79f53237eb15af304384ef50fa8d3d8ee9
          minLength: 1
          pattern: "^[0-9a-fA-F]+$"
        body:
          type: string
          description: Text der Antwort (GitLab Note).
          example: "Erledigt, habe ich angepasst."
          minLength: 1
          pattern: "^[^\\p{Cntrl}]+$"
        resolved:
          type: boolean
          description: "Wenn true, wird der Thread nach dem Antworten als resolved\
            \ markiert."
          example: true
        repo:
          type: string
          description: GitLab Projekt-ID (Repository).
          example: "12345"
          minLength: 1
          pattern: "^[0-9]+$"
      required:
      - body
      - repo
      - threadId
...
    InputFile:
      type: object
      properties:
        name:
          type: string
        lastModified:
          type: string
          format: date-time
    TemplateJsonItem:
      type: object
      properties:
        name:
          type: string
        json:
          type: string
    TemplateMdItem:
      type: object
      properties:
        canonicalUrl:
          type: string
          writeOnly: true
        markdown:
          type: string
        version:
          type: string
        url:
          type: string
    WorkspaceDetails:
      type: object
      description: "Workspace content (templates, metadata, changelog, and editor\
        \ notes) for a terminology package in a branch"
      properties:
        templatesJson:
          type: array
          items:
            $ref: "#/components/schemas/TemplateJsonItem"
          writeOnly: true
        templatesMd:
          type: array
          items:
            $ref: "#/components/schemas/TemplateMdItem"
          writeOnly: true
        packageTemplateJson:
          type: string
          description: Content of package.template.json (input/templates/package/package.template.json)
          example: "{ \"url\": \"...\", ... }"
        metadatenJson:
          type: string
          writeOnly: true
        changelogsJson:
          type: string
          description: Content of changelog.json (metadata/changelog.json)
        externalSourcesMd:
          type: string
          description: Markdown content for externalSources.md (metadata/website/package/externalSources.md)
        fhirConversionNotesMd:
          type: string
          description: Markdown content for fhirConversionNotes.md (metadata/website/package/fhirConversionNotes.md)
        noteOnAuthorMd:
          type: string
          description: Markdown content for noteOnAuthor.md (metadata/website/package/noteOnAuthor.md)
        notesOnUpdateCyclesMd:
          type: string
          description: Markdown content for notesOnUpdateCycles.md (metadata/website/package/notesOnUpdateCycles.md)
        downloadConditionsXml:
          type: string
          description: Markdown content for descriptionGeneric.md (metadata/download-conditions.md
        descriptionGenericMd:
          type: string
          description: Markdown content for descriptionGeneric.md (metadata/website/package/descriptionGeneric.md)
        inputFiles:
          type: array
          items:
            $ref: "#/components/schemas/InputFile"
          writeOnly: true
        mergeRequest:
          $ref: "#/components/schemas/MergeRequestRef"
          writeOnly: true
        templateJsons:
          type: array
          description: List of template files in the workspace (input/templates/*.template.json)
          items:
            $ref: "#/components/schemas/TemplateJsonItem"
        templateMds:
          type: array
          description: List of template markdowns metadata/website/resources-description/*.md
            via index.json
          items:
            $ref: "#/components/schemas/TemplateMdItem"
        metadataJson:
          type: string
          description: Content of metadata_package.json (metadata/metadata_package.json)
        inputFileList:
          type: array
          description: Input TFC files (input/*)
          items:
            $ref: "#/components/schemas/InputFile"
        mergeRequestRef:
          oneOf:
          - $ref: "#/components/schemas/MergeRequestRef"
          - type: "null"
    CommentItem:
      type: object
      properties:
        id:
          type: string
        type:
          type:
          - string
          - "null"
          description: Typ der Änderung/Resource. Serialisiert als String gemäß `value`.
          enum:
          - template
          - template_markdown
          - package_template
          - metadata_package
          - changelogs
          - package_markdown
          - update_index_url_version
          - download_conditions
          - input_file
        fileName:
          type:
          - string
          - "null"
        line:
          type:
          - integer
          - "null"
          format: int32
        body:
          type: string
        author:
          type:
          - string
          - "null"
        createdAt:
          type: string
          format: date-time
        resolved:
          type: boolean
        threadId:
          type:
          - string
          - "null"
        isInline:
          type: boolean
          writeOnly: true
        inline:
          type: boolean
    CommentsPayload:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentItem"
...
    ProjectItem:
      type: object
      description: A GitLab project entry
      properties:
        projectId:
          type: integer
          format: int64
          description: Internal GitLab project ID
          example: 1278
        title:
          type: string
          description: 'Project title '
          example: OPS Terminologies for BfArM
        lastModified:
          type: string
          format: date-time
          description: Timestamp of last change
          example: 2024-10-31T14:23:00Z
        description:
          type: string
          description: Description of GitLab repository
          example: Example Project
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: GitLab access token