# src/main/kotlin/de/gematik/zts/templateeditor/web/workspaces/comments/WorkspacesCommentsController.kt
/*
 * Copyright (Change Date see Readme), gematik GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * ******
 *
 * For additional notes and disclaimer from gematik and in case of changes
 * by gematik, find details in the "Readme" file.
 */

package de.gematik.zts.templateeditor.web.workspaces.comments

import de.gematik.zts.templateeditor.domain.reviews.CommentCreateRequest
import de.gematik.zts.templateeditor.domain.reviews.CommentCreateResponse
import de.gematik.zts.templateeditor.domain.reviews.CommentsPayload
import de.gematik.zts.templateeditor.domain.workspaces.ReplyRequest
import de.gematik.zts.templateeditor.gitlab.reviews.ReviewsService
import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity
import org.springframework.validation.annotation.Validated
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.RestController
import reactor.core.publisher.Mono

@RestController
@RequestMapping("/workspaces/comments")
@Validated
class WorkspacesCommentsController(
    private val reviewsService: ReviewsService,
) : WorkspacesCommentsApi {
    @GetMapping
    override fun getComments(
        @RequestParam repositoryId: String,
        @RequestParam mrId: Int,
        @RequestParam branch: String,
        @RequestParam version: String,
    ): Mono<CommentsPayload> =
        reviewsService.listComments(
            repo = repositoryId,
            mrId = mrId.toString(),
            branch = branch,
            version = version,
        )

    @PostMapping
    override fun createComment(
        @RequestBody req: CommentCreateRequest,
    ): Mono<ResponseEntity<CommentCreateResponse>> =
        reviewsService
            .createComment(req)
            .map { ResponseEntity.status(HttpStatus.CREATED).body(it) }

    @PostMapping("/reply")
    override fun reply(
        @RequestBody req: ReplyRequest,
    ): Mono<ResponseEntity<CommentCreateResponse>> =
        reviewsService
            .reply(
                repo = req.repo,
                mrId = req.mrId.toString(),
                threadId = req.threadId,
                body = req.body,
                resolved = req.resolved,
            ).map { ResponseEntity.status(HttpStatus.CREATED).body(it) }
}